cifar10: experiment with 1x1 conv layers
Taking the CIFAR10 tutorial as a base, we build a slightly different model with
the same first convolutional layer, and two final fully connected layers, but
replacing the middle convolutional layer by a 'sandwich' layer using 1x1
convolutions.

The tutorial uses a 5x5 kernel for the first convolutional layer to
capture spatial features, with 64 filters.

The second layer keeps the same kernel size and output, leading to a 5x5x64x64
layer size.

We try here to replace it with 'sandwich' layers, to reduce the depth
when the 5x5 kernel is applied:
 5x5x64x64 -> 1x1x64x32/5x5x32x32/1x1x32x64

The assumption (to be verified) is that it should reduce computation cost, yet
not degrading performance since we preserve spatial dimensions and add two
more ReLU filters.

This first model achieves 70-71% accuracy in 10000 iterations, and 73.25 % after
100000 which is lower than the tutorial version.

The processing time is equivalent on a CPU-only setup, because the
decrease in computation time coming from the reduce dimensionality in
the sandwich layer seems to be counterbalanced by the cost of the 1x1
conv layers.

To summarize, this is a rather disappointing result.