cifar10: experiment with Alex-like models
Alex Krizhevsky's work in the years 2009-2012 created the first modern
paradigm for ConvNets: a few conv layers followed by a few FC layers.

The tensorflow CIFAR10 tutorial is very Alex-like in that respect. This
directory will contain models that try different variations around that
model, trying to obtain the best compromise between accuracy and
processing power.

More specifically, the goal is to reach the same accuracy:
- 81% after 10,000 iterations with data augmentation.
... but with less parameters and operations than the tutorial model.