cifar10/nin: experiment with NIN models
Experiment with Network in Network models.

The first model is a two-layers NIN similar to the one described in the
NIN paper (although this one had three layers).

Since average along the channel dimension is not supported natively by
Tensorflow, I used the net last 1x1 conv layer to force an output of 10
features that act as 'class confidence' maps.