cifar10/data: support variable batch size
The previous version used a static batch size in the model, which
implied that the SAME batch size had to be used when a model was
saved and reloaded from disk.

We now use a dynamic batch size in the model, which allows us for
instance to perform an inference with a single image.