cifar10: add support for batch normalization
This seems to work, however the inference time is really slow at
evaluation time even if I set the training placeholder to False in the feed
dictionary.
Strangely, if I force training=False, the results are MUCH better ...

Looking at the code, I spotted a section that is not properly
conditioned to the training phase only:

in tensorflow/python/layers/normalization.py:

"Some of the computations here are not necessary when training==False
but not a constant. However, this makes the code simpler."

Wow ! Looks like someone got lazy ... or had quite a lot of GPU power
available.