cifar10: save evaluation model
When using batch normalization, the training model is not fully
optimized due to the batch normalization implementation that adds extra
ops when the 'training' value is dynamically set to 'False'.

The idea is to recreate a graph with 'training' statically set to False
at evaluation time, and freeze this model instead of the training phase
model.