snn/masquelier: cleanup and LTP bug fix
A lot of cleanup and text, and fix a bug that wrongly rewarded older
spikes.

The STDP is stil inefficient, but this time leading to a weakening of
the weights preventing the neuron from spiking.

The reward plots reveals that they are not exactly aligned with the
actual spikes, due to the late detection of these. The LIF algorithm
should perhaps be revisited to be more accurate.