snn/Masquelier: modify LTD ratio to obtain convergence
The values provided by the paper lead to the neuron saturation.
By increasing the LTD factor a little, this somehow stabilizes the
the learning for the considered input rate (64 Hz).
The whole STDP algorithm is however highly unstable, and any changes
into the input rates or learning ratios make it collapse, with the
neuron ending either fully depressed or fully activated.